reverted swipe modifier to original
