--- apps/multiplatform/common/src/desktopMain/kotlin/chat/simplex/common/views/helpers/Modifiers.desktop.kt
@@ -1,14 +1,0 @@
-package chat.simplex.common.views.helpers
-
-import androidx.compose.material.*
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-actual fun SwipeToDismissModifier(
-  state: DismissState,
-  directions: Set<DismissDirection>,
-  swipeDistance: Float,
-): Modifier {
-  return Modifier
-}
